requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==23.2
django-otp==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.7
PyJWT==2.10.1
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2025.2
qrcode==7.4
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
